requirements.txt
certifi==2020.4.5.2
chardet==3.0.4
idna==2.9
pkg-resources==0.0.0
python-chess==0.31.2
requests==2.23.0
urllib3==1.25.9
